requirements.txt
torch>=2.0.0
numpy
pillow
transformers
diffusers
bitsandbytes